Set up customization for the patterns list table
* Move flags list table customizations to a separate file
* Add columns for patterns taxonomies
* Add column for pattern flags
